+++ flow/include/flow/Arena.h
@@ -31,6 +31,7 @@
 #include "flow/Trace.h"
 #include "flow/ObjectSerializerTraits.h"
 #include "flow/FileIdentifier.h"
+#include "flow/Optional.h"
 #include <algorithm>
 #include <boost/functional/hash.hpp>
 #include <stdint.h>
@@ -216,221 +217,6 @@ inline void save(Archive& ar, const Arena& p) {
 	// No action required
 }
 
-// Optional is a wrapper for std::optional. There
-// are two primary reasons to use this wrapper instead
-// of using std::optional directly:
-//
-// 1) Legacy: A lot of code was written using Optional before
-//    std::optional was available.
-// 2) When you call get but no value is present Optional gives an
-//    assertion failure. std::optional, on the other hand, would
-//    throw std::bad_optional_access. It is easier to debug assertion
-//    failures, and FDB generally does not handle std exceptions, so
-//    assertion failures are preferable. This is the main reason we
-//    don't intend to use std::optional directly.
-template <class T>
-class Optional : public ComposedIdentifier<T, 4> {
-public:
-	using ValueType = T;
-
-	Optional() = default;
-
-	template <class U>
-	Optional(const U& t) : impl(std::in_place, t) {}
-	Optional(T&& t) : impl(std::in_place, std::move(t)) {}
-
-	/* This conversion constructor was nice, but combined with the prior constructor it means that Optional<int> can be
-	converted to Optional<Optional<int>> in the wrong way (a non-present Optional<int> converts to a non-present
-	Optional<Optional<int>>). Use .castTo<>() instead. template <class S> Optional(const Optional<S>& o) :
-	valid(o.present()) { if (valid) new (&value) T(o.get()); } */
-
-	Optional(Arena& a, const Optional<T>& o) {
-		if (o.present())
-			impl = std::make_optional<T>(a, o.get());
-	}
-	int expectedSize() const { return present() ? get().expectedSize() : 0; }
-
-	template <class R>
-	Optional<R> castTo() const {
-		return map([](const T& v) { return (R)v; });
-	}
-
-private:
-	template <class F>
-	using MapRet = std::decay_t<std::invoke_result_t<F, T>>;
-
-	template <class F>
-	using EnableIfNotMemberPointer =
-	    std::enable_if_t<!std::is_member_object_pointer_v<F> && !std::is_member_function_pointer_v<F>>;
-
-public:
-	// If the optional is set, calls the function f on the value and returns the value. Otherwise, returns an empty
-	// optional.
-	template <class F, typename = EnableIfNotMemberPointer<F>>
-	Optional<MapRet<F>> map(const F& f) const& {
-		return present() ? Optional<MapRet<F>>(f(get())) : Optional<MapRet<F>>();
-	}
-	template <class F, typename = EnableIfNotMemberPointer<F>>
-	Optional<MapRet<F>> map(const F& f) && {
-		return present() ? Optional<MapRet<F>>(f(std::move(*this).get())) : Optional<MapRet<F>>();
-	}
-
-	// Converts an Optional<T> to an Optional<R> of one of its value's members
-	//
-	// v.map(&T::member) is equivalent to v.map([](T v) { return v.member; })
-	template <class R, class Rp = std::decay_t<R>>
-	std::enable_if_t<std::is_class_v<T>, Optional<Rp>> map(
-	    R std::conditional_t<std::is_class_v<T>, T, Void>::*member) const& {
-		return present() ? Optional<Rp>(get().*member) : Optional<Rp>();
-	}
-	template <class R, class Rp = std::decay_t<R>>
-	std::enable_if_t<std::is_class_v<T>, Optional<Rp>> map(
-	    R std::conditional_t<std::is_class_v<T>, T, Void>::*member) && {
-		return present() ? Optional<Rp>(std::move(*this).get().*member) : Optional<Rp>();
-	}
-
-	// Converts an Optional<T> to an Optional<R> of a value returned by a member function of T
-	//
-	// v.map(&T::memberFunc, arg1, arg2, ...) is equivalent to
-	// v.map([](T v) { return v.memberFunc(arg1, arg2, ...); })
-	template <class R, class... Args, class Rp = std::decay_t<R>>
-	std::enable_if_t<std::is_class_v<T>, Optional<Rp>> map(
-	    R (std::conditional_t<std::is_class_v<T>, T, Void>::*memberFunc)(Args...) const,
-	    Args&&... args) const& {
-		return present() ? Optional<Rp>((get().*memberFunc)(std::forward<Args>(args)...)) : Optional<Rp>();
-	}
-	template <class R, class... Args, class Rp = std::decay_t<R>>
-	std::enable_if_t<std::is_class_v<T>, Optional<Rp>> map(
-	    R (std::conditional_t<std::is_class_v<T>, T, Void>::*memberFunc)(Args...) const,
-	    Args&&... args) && {
-		return present() ? Optional<Rp>((std::move(*this).get().*memberFunc)(std::forward<Args>(args)...))
-		                 : Optional<Rp>();
-	}
-
-	// Given T that is a pointer or pointer-like type to type P (e.g. T=P* or T=Reference<P>), converts an Optional<T>
-	// to an Optional<R> of one its value's members. If the optional value is present and false-like (null), then
-	// returns an empty Optional<R>.
-	//
-	// v.mapRef(&P::member) is equivalent to Optional<R>(v.get()->member) if v is present and non-null
-	template <class P, class R, class Rp = std::decay_t<R>>
-	std::enable_if_t<std::is_class_v<T> || std::is_pointer_v<T>, Optional<Rp>> mapRef(R P::*member) const& {
-		if (!present() || !get()) {
-			return Optional<Rp>();
-		}
-
-		P& p = *get();
-		return p.*member;
-	}
-
-	// Given T that is a pointer or pointer-like type to type P (e.g. T=P* or T=Reference<P>), converts an Optional<T>
-	// to an Optional<R> of a value returned by a member function of P. If the optional value is present and false-like
-	// (null), then returns an empty Optional<R>.
-	//
-	// v.mapRef(&T::memberFunc, arg1, arg2, ...) is equivalent to Optional<R>(v.get()->memberFunc(arg1, arg2, ...)) if v
-	// is present and non-null
-	template <class P, class R, class... Args, class Rp = std::decay_t<R>>
-	std::enable_if_t<std::is_class_v<T> || std::is_pointer_v<T>, Optional<Rp>> mapRef(R (P::*memberFunc)(Args...) const,
-	                                                                                  Args&&... args) const& {
-		if (!present() || !get()) {
-			return Optional<Rp>();
-		}
-		P& p = *get();
-		return (p.*memberFunc)(std::forward<Args>(args)...);
-	}
-
-	// Similar to map with a mapped type of Optional<R>, but flattens the result. For example, if the mapped result is
-	// of type Optional<R>, map will return Optional<Optional<R>> while flatMap will return Optional<R>
-	template <class... Args>
-	auto flatMap(Args&&... args) const& {
-		auto val = map(std::forward<Args>(args)...);
-		using R = typename decltype(val)::ValueType::ValueType;
-
-		if (val.present()) {
-			return val.get();
-		} else {
-			return Optional<R>();
-		}
-	}
-	template <class... Args>
-	auto flatMap(Args&&... args) && {
-		auto val = std::move(*this).map(std::forward<Args>(args)...);
-		using R = typename decltype(val)::ValueType::ValueType;
-
-		if (val.present()) {
-			return val.get();
-		} else {
-			return Optional<R>();
-		}
-	}
-
-	// Similar to mapRef with a mapped type of Optional<R>, but flattens the result. For example, if the mapped result
-	// is of type Optional<R>, mapRef will return Optional<Optional<R>> while flatMapRef will return Optional<R>
-	template <class... Args>
-	auto flatMapRef(Args&&... args) const& {
-		auto val = mapRef(std::forward<Args>(args)...);
-		using R = typename decltype(val)::ValueType::ValueType;
-
-		if (val.present()) {
-			return val.get();
-		} else {
-			return Optional<R>();
-		}
-	}
-	template <class... Args>
-	auto flatMapRef(Args&&... args) && {
-		auto val = std::move(*this).mapRef(std::forward<Args>(args)...);
-		using R = typename decltype(val)::ValueType::ValueType;
-
-		if (val.present()) {
-			return val.get();
-		} else {
-			return Optional<R>();
-		}
-	}
-
-	bool present() const { return impl.has_value(); }
-	T& get() & {
-		UNSTOPPABLE_ASSERT(impl.has_value());
-		return impl.value();
-	}
-	T const& get() const& {
-		UNSTOPPABLE_ASSERT(impl.has_value());
-		return impl.value();
-	}
-	T&& get() && {
-		UNSTOPPABLE_ASSERT(impl.has_value());
-		return std::move(impl.value());
-	}
-	template <class U>
-	T orDefault(U&& defaultValue) const& {
-		return impl.value_or(std::forward<U>(defaultValue));
-	}
-	template <class U>
-	T orDefault(U&& defaultValue) && {
-		return std::move(impl).value_or(std::forward<U>(defaultValue));
-	}
-
-	// Spaceship operator.  Treats not-present as less-than present.
-	int compare(Optional const& rhs) const {
-		if (present() == rhs.present()) {
-			return present() ? get().compare(rhs.get()) : 0;
-		}
-		return present() ? 1 : -1;
-	}
-
-	bool operator==(Optional const& o) const { return impl == o.impl; }
-	bool operator!=(Optional const& o) const { return !(*this == o); }
-	// Ordering: If T is ordered, then Optional() < Optional(t) and (Optional(u)<Optional(v))==(u<v)
-	bool operator<(Optional const& o) const { return impl < o.impl; }
-
-	void reset() { impl.reset(); }
-	size_t hash() const { return hashFunc(impl); }
-
-private:
-	static inline std::hash<std::optional<T>> hashFunc{};
-	std::optional<T> impl;
-};
-
 template <class Archive, class T>
 inline void load(Archive& ar, Optional<T>& value) {
 	bool valid;
@@ -569,7 +355,7 @@ class StringRef {
 public:
 	constexpr static FileIdentifier file_identifier = 13300811;
 	StringRef() : data(0), length(0) {}
-	StringRef(Arena& p, const StringRef& toCopy) : data(new(p) uint8_t[toCopy.size()]), length(toCopy.size()) {
+	StringRef(Arena& p, const StringRef& toCopy) : data(new (p) uint8_t[toCopy.size()]), length(toCopy.size()) {
 		if (length > 0) {
 			memcpy((void*)data, toCopy.data, length);
 		}
@@ -580,7 +366,7 @@ public:
 		if (length)
 			memcpy((void*)data, &toCopy[0], length);
 	}
-	StringRef(Arena& p, const uint8_t* toCopy, int length) : data(new(p) uint8_t[length]), length(length) {
+	StringRef(Arena& p, const uint8_t* toCopy, int length) : data(new (p) uint8_t[length]), length(length) {
 		if (length > 0) {
 			memcpy((void*)data, toCopy, length);
 		}
@@ -1133,7 +919,7 @@ public:
 	// Arena constructor for non-Ref types, identified by !flow_ref
 	template <class T2 = T, VecSerStrategy S>
 	VectorRef(Arena& p, const VectorRef<T, S>& toCopy, typename std::enable_if<!flow_ref<T2>::value, int>::type = 0)
-	  : VPS(toCopy), data((T*)new(p) uint8_t[sizeof(T) * toCopy.size()]), m_size(toCopy.size()),
+	  : VPS(toCopy), data((T*)new (p) uint8_t[sizeof(T) * toCopy.size()]), m_size(toCopy.size()),
 	    m_capacity(toCopy.size()) {
 		if (m_size > 0) {
 			std::copy(toCopy.data, toCopy.data + m_size, data);
@@ -1143,7 +929,7 @@ public:
 	// Arena constructor for Ref types, which must have an Arena constructor
 	template <class T2 = T, VecSerStrategy S>
 	VectorRef(Arena& p, const VectorRef<T, S>& toCopy, typename std::enable_if<flow_ref<T2>::value, int>::type = 0)
-	  : VPS(), data((T*)new(p) uint8_t[sizeof(T) * toCopy.size()]), m_size(toCopy.size()), m_capacity(toCopy.size()) {
+	  : VPS(), data((T*)new (p) uint8_t[sizeof(T) * toCopy.size()]), m_size(toCopy.size()), m_capacity(toCopy.size()) {
 		for (int i = 0; i < m_size; i++) {
 			auto ptr = new (&data[i]) T(p, toCopy[i]);
 			VPS::add(*ptr);
